# docs/swagger.yml
openapi: 3.0.1
info:
  title: E-commerce API
  description: API documentation for the E-commerce application
  version: 1.0.0
servers:
  - url: http://localhost:8080
paths:
  /api/products:
    get:
      summary: Get all products
      tags: 
        - Products
      parameters:
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            default: 10
          description: Limit the number of products returned
        - name: page
          in: query
          required: false
          schema:
            type: integer
            default: 1
          description: The page number for pagination
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Product'
        '404':
          description: No products found
        '500':
          description: Internal server error
...
  /api/users:
    get:
      summary: Get all users (Admin only)
      tags: 
        - Users
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/User'
        '500':
          description: Internal server error

    post:
      summary: Register a new user
      tags: 
        - Users
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/User'
      responses:
        '201':
          description: User registered successfully
        '400':
          description: Invalid input
        '500':
          description: Internal server error
...
components:
  schemas:
    Product:
      type: object
      properties:
        id:
          type: string
        title:
          type: string
        description:
          type: string
        price:
          type: number
        stock:
          type: integer
      example:
        id: "607c191e810c19729de860ea"
        title: "Product A"
        description: "Description of Product A"
        price: 100
        stock: 10
...
    User:
      type: object
      properties:
        id:
          type: string
        first_name:
          type: string
        last_name:
          type: string
        email:
          type: string
        role:
          type: string
      example:
        id: "609c1fd5810c19729de8612d"
        first_name: "John"
        last_name: "Doe"
        email: "john.doe@example.com"
        role: "user"
...